+++ fs/inode/Make.defs
@@ -36,7 +36,7 @@
 CSRCS += fs_files.c fs_foreachinode.c fs_inode.c fs_inodeaddref.c
 CSRCS += fs_inodebasename.c fs_inodefind.c fs_inodefree.c fs_inoderelease.c
 CSRCS += fs_inoderemove.c fs_inodereserve.c fs_inodesearch.c
-CSRCS += fs_fileopen.c fs_filedetach.c fs_fileclose.c
+CSRCS += fs_fileopen.c fs_filedetach.c
 
 # Include inode/utils build support
